replace as.numeric with as.integer in tags.to.onehot
Former-commit-id: 88319c17fdaf4a9d421875fbd63ee66c07a9ddc6

+++ R/tagCell.R
@@ -224,7 +224,7 @@ tags.to.onehot <- function(tags.df, exclude.cols = c("pos", "cellID", "viewed"))
     mutate(ucid = as.character.factor(ucid),
            t.frame = as.character.factor(t.frame)) %>%
     # Then convert everything to numeric type
-    mutate_all(as.numeric)
+    mutate_all(as.integer)
   
   return(annotations.dt)
 }
